chore: engine type options order
